consolidate hpt() and host2ip() into utils.c
